skills-optional/word-processor/requirements.txt
python-docx>=0.8.11
Pillow>=9.0.0
pypandoc-binary>=1.11  # 可选，用于格式转换
